Router Refresh command and Fixes
